+++ app/src/main/java/woowacourse/shopping/data/httpclient/mapper/OrderResponseMapper.kt
@@ -1,13 +1,13 @@
-package woowacourse.shopping.data.remote.mapper
+package woowacourse.shopping.data.httpclient.mapper
 
+import woowacourse.shopping.data.httpclient.response.order.Individualorder.IndividualOrderResponse
+import woowacourse.shopping.data.httpclient.response.order.Individualorder.OrderItem
 import woowacourse.shopping.data.model.DataCount
 import woowacourse.shopping.data.model.DataOrder
 import woowacourse.shopping.data.model.DataOrderItem
 import woowacourse.shopping.data.model.DataPoint
 import woowacourse.shopping.data.model.DataPrice
 import woowacourse.shopping.data.model.DataProduct
-import woowacourse.shopping.data.remote.response.order.Individualorder.IndividualOrderResponse
-import woowacourse.shopping.data.remote.response.order.Individualorder.OrderItem
 import woowacourse.shopping.util.LocalDateTimeFormatter
 import java.time.LocalDateTime
 
@@ -25,7 +25,10 @@ fun OrderItem.toData(): DataOrderItem =
 fun IndividualOrderResponse.toData(): DataOrder =
     DataOrder(
         orderId = orderId,
-        createdAt = LocalDateTime.parse(createdAt, LocalDateTimeFormatter.isoLocalDateTimeFormatter),
+        createdAt = LocalDateTime.parse(
+            createdAt,
+            LocalDateTimeFormatter.isoLocalDateTimeFormatter
+        ),
         orderItems = orderItems.map { it.toData() },
         totalPrice = DataPrice(totalPrice),
         usedPoint = DataPoint(usedPoint),
