+++ app/src/main/java/woowacourse/shopping/data/httpclient/request/AddOrderRequest.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping.data.remote.request
+package woowacourse.shopping.data.httpclient.request
 
 import com.google.gson.annotations.SerializedName
