+++ app/src/main/java/woowacourse/shopping/data/httpclient/request/BasketUpdateRequest.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping.data.remote.request
+package woowacourse.shopping.data.httpclient.request
 
 data class BasketUpdateRequest(
     val quantity: Int
