+++ app/src/main/java/woowacourse/shopping/data/httpclient/service/BasketService.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping.data.remote.service
+package woowacourse.shopping.data.httpclient.service
 
 import retrofit2.Call
 import retrofit2.http.Body
@@ -7,9 +7,9 @@ import retrofit2.http.GET
 import retrofit2.http.PATCH
 import retrofit2.http.POST
 import retrofit2.http.Path
+import woowacourse.shopping.data.httpclient.request.BasketAddRequest
+import woowacourse.shopping.data.httpclient.request.BasketUpdateRequest
 import woowacourse.shopping.data.model.DataBasketProduct
-import woowacourse.shopping.data.remote.request.BasketAddRequest
-import woowacourse.shopping.data.remote.request.BasketUpdateRequest
 
 interface BasketService {
     @GET("/cart-items")
