+++ app/src/main/java/woowacourse/shopping/data/httpclient/service/ProductService.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping.data.remote.service
+package woowacourse.shopping.data.httpclient.service
 
 import retrofit2.Call
 import retrofit2.http.GET
