+++ app/src/main/java/woowacourse/shopping/data/httpclient/service/UserPointInfoService.kt
@@ -1,8 +1,8 @@
-package woowacourse.shopping.data.remote.service
+package woowacourse.shopping.data.httpclient.service
 
 import retrofit2.Call
 import retrofit2.http.GET
-import woowacourse.shopping.data.remote.response.UserPointInfoResponse
+import woowacourse.shopping.data.httpclient.response.UserPointInfoResponse
 
 interface UserPointInfoService {
     @GET("/users")
